refactor: Behalte Einheiten in gesamter Bibliothek bei

+++ src/Interfaces/Position.php
@@ -10,11 +10,9 @@ interface Position extends \JsonSerializable, Formatable
 
     public function quantity(): float;
 
-    // Can be a fraction of the smallest currency unit
     public function price(): float;
 
-    // Representas the smallest currency unit and must therfor be a an interger
-    public function amount(): int;
+    public function amount(): float;
 
     public function format(string $method): string;
